Move helpers to lib/ApiClient object
